requirements.txt
requests==2.27.1
celery==5.2.3
kombu==5.2.3
stripe==2.66.0
proxmoxer==1.2.0
python-dateutil==2.8.2
flower==1.0.0
gevent==21.12.0
gunicorn==20.1.0
supervisor==4.2.4
sentry-sdk==1.5.5
celery-singleton==0.3.1
certbot==1.23.0
certbot-nginx==1.23.0
nginx-config-builder==1.0.1
psycopg2==2.9.3
Django<4
djangorestframework==3.13.1
dj-stripe==2.6.1
drf-aggregates==0.0.13
dj-rest-auth==2.2.3
django-allauth==0.48.0
django-celery-beat==2.1.0
django-celery-results==2.2.0
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-filter==21.1
djangorestframework-datatables==0.7.0
drf-schema-adapter==2.1.1
